Move jwt cookie and token generation to it's own package rather than utils
